requirements.txt
# Core
fastapi>=0.95.2
uvicorn>=0.24.0

# Async PostgreSQL
asyncpg>=0.27.0

# Kafka
aiokafka>=0.8.0

# Data validation
pydantic>=1.10.7,<2.0.0

# Async environment
python-multipart>=0.0.6  # Required by FastAPI
python-dotenv>=0.19.0    # For .env file support

# Optional but recommended for production
uvloop>=0.17.0           # Faster event loop for Linux
httptools>=0.5.0         # HTTP parser